changed DOS contract to a slightly more secure version where retrieved from address needs to be equal the new from parameter of the method
when the signature and the input data does not match, Utils.recover resp. Solidity's ecrecover returns another (random) address. If this address has tokens, the tokens would be transferred. Therefore we pass the from address as parameter and check if the retrieved from is the same as the passed one.

@tbocek